rental refactor
