+++ app.rb
@@ -4,6 +4,7 @@ require_relative './book'
 require_relative './rental'
 require_relative './person_module'
 require_relative './book_module'
+require_relative './rental_module'
 
 class App
   def initialize(option)
@@ -15,6 +16,7 @@ class App
   
   include PersonModule
   include BookModule
+  include RentalModule
 
   # List all books.
   def list_books
@@ -29,27 +31,6 @@ class App
     @persons.each do |person|
       puts "[#{person.class.name}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
     end
-    @option.show_options
-  end
-
-  # Create a rental.
-  def create_rental
-    puts 'Select a book from the following list by number'
-    @books.each_with_index do |book, index|
-      puts "#{index} Title: '#{book.title}', Author: #{book.author}"
-    end
-    book_number = gets.chomp.to_i
-    puts 'Select a person from the following list by number'
-    @persons.each_with_index do |person, index|
-      puts "#{index} [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    person_number = gets.chomp.to_i
-
-    print 'Date: '
-    date = gets.chomp
-    @rentals.push(Rental.new(date, @books[book_number], @persons[person_number]))
-    puts 'Rental created successfully'
-
     @option.show_options
   end
