+++ rental_module.rb
@@ -1,0 +1,21 @@
+module RentalModule
+    def create_rental
+        puts 'Select a book from the following list by number'
+        @books.each_with_index do |book, index|
+          puts "#{index} Title: '#{book.title}', Author: #{book.author}"
+        end
+        book_number = gets.chomp.to_i
+        puts 'Select a person from the following list by number'
+        @persons.each_with_index do |person, index|
+          puts "#{index} [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+        end
+        person_number = gets.chomp.to_i
+    
+        print 'Date: '
+        date = gets.chomp
+        @rentals.push(Rental.new(date, @books[book_number], @persons[person_number]))
+        puts 'Rental created successfully'
+    
+        @option.show_options
+      end    
+end
